deleted edited

+++ src/Pages/Admin/MyToys.jsx
@@ -22,23 +22,35 @@ const MyToys = () => {
     }, [user, value])
 
     const handleDelete = id =>{
-        fetch(`https://educational-toys-server.vercel.app/allToys/${id}`, {
-            method: 'DELETE'
-        })
-        .then(res =>res.json())
-        .then(data =>{
-            console.log(data)
-            if(data.deletedCount > 0){
-                Swal.fire({
-                    title: 'Delete!',
-                    text: 'Sorry! You have deleted one toy.',
-                    icon: 'success',
-                    confirmButtonText: 'Cool'
-                  })
-                  const remaining = toys.filter(toy => toy._id !== id);
-                  setToys(remaining);
-            }
-        })
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            fetch(`https://educational-toys-server.vercel.app/allToys/${id}`, {
+                method: 'DELETE'
+            })
+            .then(res =>res.json())
+            .then(data =>{
+                console.log(data)
+                if(data.deletedCount > 0){
+                    Swal.fire({
+                        title: 'Delete!',
+                        text: 'Sorry! You have deleted one toy.',
+                        icon: 'success',
+                        confirmButtonText: 'Cool'
+                      })
+                      const remaining = toys.filter(toy => toy._id !== id);
+                      setToys(remaining);
+                }
+            })
+          })
+
+        
     }
     return (
         <div className='mb-10 bg-green-700 rounded-xl p-10'>
